Added data formatter files to where needed
